Промених в базата данни типа данни на измерваните величини от  float на double.
Добавих в проекта Models папка DTO.
Там съм създал моделите, които практически ще трябва да се показват в таблицата в графичният екран. За всяка една величина има по две таблици - с всички измерени стойности и само със излезлите от нормални граници.

Предполагам, че с Data  Binding-а ще трябва да се привърже DTO таблицата, със графичната контрола Table.

Утре ще има запълнени стойности от измервания, тогава ще напиша методите и за DTO таблиците.

+++ Excavators.Models/TempSensorType.cs
@@ -23,8 +23,8 @@
         public string Description { get; set; }
 
 
-        public float RangeMin { get; set; }
-        public float RangeMax { get; set; }
+        public double RangeMin { get; set; }
+        public double RangeMax { get; set; }
 
 
         public virtual ICollection<TempSensor> TempSensors { get; set; }
